Add clean-code
