+++ SettingScreen/Sources/App/SceneDelegate.swift
@@ -16,12 +16,10 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
                options connectionOptions: UIScene.ConnectionOptions) {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         let window = UIWindow(windowScene: windowScene)
-//        let model = Models()
-        let viewController = ViewController()
+        let viewController = SettingScreenViewController()
         let navigatorController = UINavigationController(rootViewController: viewController)
         window.rootViewController = navigatorController
         window.makeKeyAndVisible()
         self.window = window
     }
 }
-
