+++ SettingScreen/Sources/Scene/View/CustomModule/SwichTableViewCell.swift
@@ -84,4 +84,3 @@ class SwichTableViewCell: UITableViewCell {
         ])
     }
 }
-
